Add files via upload
This is the updated version with the new cladogram picture and the mappings for it. There are 12 areas mapped, one for each of the categories. Right now, if a mapped area is clicked on a number is passed to the test method, and an alert is displayed based on which number is passed in. I also included a word document that has the ordering of the land plants and the numbers that are used with them in the mapping(1-12). I figure we can use it as a reference.